fix(dsp): close C6 unity-exit click and block-size PDC error; zero warnings
Unity path: the STFT pipeline now runs every non-bypassed block; at settled
unity the output is overwritten from the latency-matched delay line
(bit-perfect passthrough kept, rings and mask stats stay warm, spectrum live
at defaults). Delay line is written every block and read at a fixed offset
behind the write pointer, making its delay structurally exact. Unity epsilon
widened 1e-8 -> 1e-4 (H15).

Latency: reported figure is now a full fftSize (2048, ~43 ms @48k) and the
output ring is primed with one window of zeros in prepare()/reset(). The old
1536 was only realised at hop-multiple host buffers; smaller buffers landed
up to 448 samples late vs PDC and non-divisor buffers (REAPER variable
blocks) glitched via ring underflow. Realised latency is now block-size
invariant; rings are sized from maxBlockSize for exotic hosts.

Also: brightness smoother advances per sample while ramping (kills the
per-block automation zipper); spectrum snapshot publishes per STFT hop
instead of per block; gain params get a 1.71-skew fader taper; all 260
-Wsign-conversion/-Wshadow warnings in the DSP and processor cleared with
zero-behavior casts (harness output byte-identical).

Harness: two new permanent gates - per-block-size latency truth
(64/128/256/333/512/1024 impulse == reported, dropout scans) and
unity->active transition continuity. All isolation targets still MET
(sine -146.4, hum -52.7, noise -54.4, clicks -117.5 dB at sep=100%).
stft-validator PASS (three runs); pluginval s8 SUCCESS; auval PASS.

+++ Source/DSP/MagPhaseFrame.cpp
@@ -171,7 +171,7 @@ float MagPhaseFrame::calculateEnergy() const noexcept
     // Use JUCE's efficient sum-of-squares calculation
     for (int i = 0; i < numBins; ++i)
     {
-        const float mag = magnitudeData_[i];
+        const float mag = magnitudeData_[static_cast<size_t>(i)];
         energy += mag * mag;
     }
